+++ derive/src/args.rs
@@ -1,14 +1,12 @@
 use std::fmt::{self, Display, Formatter};
 
 use darling::{
-    ast::{Data, Fields},
+    ast::{Data, Fields, NestedMeta},
     util::{Ignored, SpannedValue},
     FromDeriveInput, FromField, FromMeta, FromVariant,
 };
 use inflector::Inflector;
-use syn::{
-    Attribute, Generics, Ident, Lit, LitBool, LitStr, Meta, NestedMeta, Path, Type, Visibility,
-};
+use syn::{Attribute, Generics, Ident, Lit, LitBool, LitStr, Meta, Path, Type, Visibility};
 
 use crate::validators::Validators;
 
@@ -526,7 +524,6 @@ pub struct InterfaceFieldArgument {
     pub name: String,
     #[darling(default)]
     pub desc: Option<String>,
-    #[darling(rename = "type")]
     pub ty: LitStr,
     #[darling(default)]
     pub default: Option<DefaultValue>,
@@ -545,7 +542,6 @@ pub struct InterfaceFieldArgument {
 #[derive(FromMeta)]
 pub struct InterfaceField {
     pub name: SpannedValue<String>,
-    #[darling(rename = "type")]
     pub ty: LitStr,
     #[darling(default)]
     pub method: Option<String>,