+++ derive/src/complex_object.rs
@@ -33,7 +33,7 @@ pub fn generate(
     // Computation of the derivated fields
     let mut derived_impls = vec![];
     for item in &mut item_impl.items {
-        if let ImplItem::Method(method) = item {
+        if let ImplItem::Fn(method) = item {
             let method_args: args::ComplexObjectField =
                 parse_graphql_attrs(&method.attrs)?.unwrap_or_default();
 
@@ -108,7 +108,7 @@ pub fn generate(
 
                     new_impl.block = syn::parse2::<Block>(new_block).expect("invalid block");
 
-                    derived_impls.push(ImplItem::Method(new_impl));
+                    derived_impls.push(ImplItem::Fn(new_impl));
                 }
             }
         }
@@ -116,7 +116,7 @@ pub fn generate(
     item_impl.items.append(&mut derived_impls);
 
     for item in &mut item_impl.items {
-        if let ImplItem::Method(method) = item {
+        if let ImplItem::Fn(method) = item {
             let method_args: args::ComplexObjectField =
                 parse_graphql_attrs(&method.attrs)?.unwrap_or_default();
             if method_args.skip {