+++ derive/src/lib.rs
@@ -27,15 +27,30 @@ mod validators;
 
 use darling::{FromDeriveInput, FromMeta};
 use proc_macro::TokenStream;
-use syn::{parse_macro_input, AttributeArgs, DeriveInput, ItemFn, ItemImpl};
+use syn::{parse_macro_input, DeriveInput, ItemFn, ItemImpl};
+
+macro_rules! parse_nested_meta {
+    ($ty:ty, $args:expr) => {{
+        let meta = match darling::ast::NestedMeta::parse_meta_list(proc_macro2::TokenStream::from(
+            $args,
+        )) {
+            Ok(v) => v,
+            Err(e) => {
+                return TokenStream::from(darling::Error::from(e).write_errors());
+            }
+        };
+
+        match <$ty>::from_list(&meta) {
+            Ok(object_args) => object_args,
+            Err(err) => return TokenStream::from(err.write_errors()),
+        }
+    }};
+}
 
 #[proc_macro_attribute]
 #[allow(non_snake_case)]
 pub fn Object(args: TokenStream, input: TokenStream) -> TokenStream {
-    let object_args = match args::Object::from_list(&parse_macro_input!(args as AttributeArgs)) {
-        Ok(object_args) => object_args,
-        Err(err) => return TokenStream::from(err.write_errors()),
-    };
+    let object_args = parse_nested_meta!(args::Object, args);
     let mut item_impl = parse_macro_input!(input as ItemImpl);
     match object::generate(&object_args, &mut item_impl) {
         Ok(expanded) => expanded,
@@ -59,11 +74,7 @@ pub fn derive_simple_object(input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 #[allow(non_snake_case)]
 pub fn ComplexObject(args: TokenStream, input: TokenStream) -> TokenStream {
-    let object_args =
-        match args::ComplexObject::from_list(&parse_macro_input!(args as AttributeArgs)) {
-            Ok(object_args) => object_args,
-            Err(err) => return TokenStream::from(err.write_errors()),
-        };
+    let object_args = parse_nested_meta!(args::ComplexObject, args);
     let mut item_impl = parse_macro_input!(input as ItemImpl);
     match complex_object::generate(&object_args, &mut item_impl) {
         Ok(expanded) => expanded,
@@ -125,11 +136,7 @@ pub fn derive_union(input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 #[allow(non_snake_case)]
 pub fn Subscription(args: TokenStream, input: TokenStream) -> TokenStream {
-    let object_args =
-        match args::Subscription::from_list(&parse_macro_input!(args as AttributeArgs)) {
-            Ok(object_args) => object_args,
-            Err(err) => return TokenStream::from(err.write_errors()),
-        };
+    let object_args = parse_nested_meta!(args::Subscription, args);
     let mut item_impl = parse_macro_input!(input as ItemImpl);
     match subscription::generate(&object_args, &mut item_impl) {
         Ok(expanded) => expanded,
@@ -140,10 +147,7 @@ pub fn Subscription(args: TokenStream, input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 #[allow(non_snake_case)]
 pub fn Scalar(args: TokenStream, input: TokenStream) -> TokenStream {
-    let scalar_args = match args::Scalar::from_list(&parse_macro_input!(args as AttributeArgs)) {
-        Ok(scalar_args) => scalar_args,
-        Err(err) => return TokenStream::from(err.write_errors()),
-    };
+    let scalar_args = parse_nested_meta!(args::Scalar, args);
     let mut item_impl = parse_macro_input!(input as ItemImpl);
     match scalar::generate(&scalar_args, &mut item_impl) {
         Ok(expanded) => expanded,
@@ -207,11 +211,7 @@ pub fn derive_newtype(input: TokenStream) -> TokenStream {
 #[proc_macro_attribute]
 #[allow(non_snake_case)]
 pub fn Directive(args: TokenStream, input: TokenStream) -> TokenStream {
-    let directive_args =
-        match args::Directive::from_list(&parse_macro_input!(args as AttributeArgs)) {
-            Ok(directive_args) => directive_args,
-            Err(err) => return TokenStream::from(err.write_errors()),
-        };
+    let directive_args = parse_nested_meta!(args::Directive, args);
     let mut item_fn = parse_macro_input!(input as ItemFn);
     match directive::generate(&directive_args, &mut item_fn) {
         Ok(expanded) => expanded,